+++ ext/stub/12__closure.zep.c
@@ -22,7 +22,8 @@ ZEPHIR_INIT_CLASS(stub_12__closure)
 {
 	ZEPHIR_REGISTER_CLASS(stub, 12__closure, stub, 12__closure, stub_12__closure_method_entry, ZEND_ACC_FINAL_CLASS);
 
-	zend_declare_property_null(stub_12__closure_ce, SL("name"), ZEND_ACC_PUBLIC|ZEND_ACC_STATIC);
+	zend_declare_property_null(stub_12__closure_ce, SL("name"), ZEND_ACC_PUBLIC);
+	zend_declare_property_null(stub_12__closure_ce, SL("__$zephir_this"), ZEND_ACC_PUBLIC);
 	return SUCCESS;
 }
 
@@ -30,14 +31,18 @@ PHP_METHOD(stub_12__closure, __invoke)
 {
 	zephir_method_globals *ZEPHIR_METHOD_GLOBALS_PTR = NULL;
 	zend_long ZEPHIR_LAST_CALL_STATUS;
-	zval name, _0;
+	zval name, __$zephir_this, _0;
 	zval *this_ptr = getThis();
 
 	ZVAL_UNDEF(&name);
+	ZVAL_UNDEF(&__$zephir_this);
 	ZVAL_UNDEF(&_0);
 	ZEPHIR_METHOD_GLOBALS_PTR = pecalloc(1, sizeof(zephir_method_globals), 0);
 	zephir_memory_grow_stack(ZEPHIR_METHOD_GLOBALS_PTR, __func__);
-	zephir_read_static_property_ce(&name, stub_12__closure_ce, SL("name"), PH_NOISY_CC);
+	zephir_read_property(&__$zephir_this, this_ptr, SL("__$zephir_this"), PH_NOISY_CC | PH_READONLY);
+	zephir_memory_observe(&name);
+	zephir_read_property(&name, this_ptr, SL("name"), PH_NOISY_CC);
+	this_ptr = &__$zephir_this;
 
 	ZEPHIR_CALL_METHOD(&_0, this_ptr, "issue2497helper", NULL, 0);
 	zephir_check_call_status();