+++ ext/stub/23__closure.zep.c
@@ -13,24 +13,30 @@
 
 #include "kernel/main.h"
 #include "kernel/object.h"
+#include "kernel/memory.h"
 
 
 ZEPHIR_INIT_CLASS(stub_23__closure)
 {
 	ZEPHIR_REGISTER_CLASS(stub, 23__closure, stub, 23__closure, stub_23__closure_method_entry, ZEND_ACC_FINAL_CLASS);
 
-	zend_declare_property_null(stub_23__closure_ce, SL("name"), ZEND_ACC_PUBLIC|ZEND_ACC_STATIC);
+	zend_declare_property_null(stub_23__closure_ce, SL("name"), ZEND_ACC_PUBLIC);
 	return SUCCESS;
 }
 
 PHP_METHOD(stub_23__closure, __invoke)
 {
+	zephir_method_globals *ZEPHIR_METHOD_GLOBALS_PTR = NULL;
 	zval name;
+	zval *this_ptr = getThis();
 
 	ZVAL_UNDEF(&name);
-	zephir_read_static_property_ce(&name, stub_23__closure_ce, SL("name"), PH_NOISY_CC);
+	ZEPHIR_METHOD_GLOBALS_PTR = pecalloc(1, sizeof(zephir_method_globals), 0);
+	zephir_memory_grow_stack(ZEPHIR_METHOD_GLOBALS_PTR, __func__);
+	zephir_memory_observe(&name);
+	zephir_read_property(&name, this_ptr, SL("name"), PH_NOISY_CC);
 
 	RETVAL_ZVAL(&name, 1, 0);
-	return;
+	RETURN_MM();
 }