+++ ext/stub/40__closure.zep.h
@@ -1,0 +1,16 @@
+
+extern zend_class_entry *stub_40__closure_ce;
+
+ZEPHIR_INIT_CLASS(stub_40__closure);
+
+PHP_METHOD(stub_40__closure, __invoke);
+
+ZEND_BEGIN_ARG_INFO_EX(arginfo_stub_40__closure___invoke, 0, 0, 2)
+	ZEND_ARG_INFO(0, carry)
+	ZEND_ARG_INFO(0, row)
+ZEND_END_ARG_INFO()
+
+ZEPHIR_INIT_FUNCS(stub_40__closure_method_entry) {
+	PHP_ME(stub_40__closure, __invoke, arginfo_stub_40__closure___invoke, ZEND_ACC_PUBLIC|ZEND_ACC_FINAL)
+	PHP_FE_END
+};