+++ ext/stub/8__closure.zep.c
@@ -20,16 +20,17 @@ ZEPHIR_INIT_CLASS(stub_8__closure)
 {
 	ZEPHIR_REGISTER_CLASS(stub, 8__closure, stub, 8__closure, stub_8__closure_method_entry, ZEND_ACC_FINAL_CLASS);
 
-	zend_declare_property_null(stub_8__closure_ce, SL("abc"), ZEND_ACC_PUBLIC|ZEND_ACC_STATIC);
+	zend_declare_property_null(stub_8__closure_ce, SL("abc"), ZEND_ACC_PUBLIC);
 	return SUCCESS;
 }
 
 PHP_METHOD(stub_8__closure, __invoke)
 {
 	zval abc;
+	zval *this_ptr = getThis();
 
 	ZVAL_UNDEF(&abc);
-	zephir_read_static_property_ce(&abc, stub_8__closure_ce, SL("abc"), PH_NOISY_CC);
+	zephir_read_property(&abc, this_ptr, SL("abc"), PH_NOISY_CC | PH_READONLY);
 
 	RETURN_LONG((zephir_get_numberval(&abc) + 1));
 }